[#14][FEATURE][REFACTOR] Timeline.vue 추가 및 폴더명 변경    - 수정이유: components와 pages 폴더 구조 미일치, EASYXPLAIN '연혁' 추가 필요    - 수정내용: pages/service/easyxplain/ 폴더 구조 반영 및 Timeline.vue 추가
